feat: model 분리
