+++ app/api/content/content_crud.py
@@ -13,7 +13,7 @@ from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm import Session
 
 from api.content.content_schema import ContentCreate
-from models import Content
+from models.content import Content
 
 
 def create_content(current_user: dict, db: Session, content_create: ContentCreate):
