+++ app/api/image/image_crud.py
@@ -14,7 +14,10 @@ from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm import Session
 
 from api.image.image_schema import ImageCreate
-from models import ContentImage, Image, User, UserImage
+from models.content_image import ContentImage
+from models.image import Image
+from models.user import User
+from models.user_image import UserImage
 
 load_dotenv()
